@@ -1,6 +1,11 @@
 package personal.carlthronson.dl.be;
 
 
+import java.io.IOException;
+import java.io.InputStream;
+import java.io.OutputStream;
+import java.util.EnumSet;
+
 import com.amazonaws.serverless.exceptions.ContainerInitializationException;
 import com.amazonaws.serverless.proxy.internal.testutils.Timer;
 import com.amazonaws.serverless.proxy.model.AwsProxyRequest;
@@ -10,13 +15,8 @@ import com.amazonaws.serverless.proxy.spring.SpringBootLambdaContainerHandler;
 import com.amazonaws.services.lambda.runtime.Context;
 import com.amazonaws.services.lambda.runtime.RequestStreamHandler;
 
-import javax.servlet.DispatcherType;
-import javax.servlet.FilterRegistration;
-
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.OutputStream;
-import java.util.EnumSet;
+import jakarta.servlet.DispatcherType;
+import jakarta.servlet.FilterRegistration;
 
 
 public class StreamLambdaHandler implements RequestStreamHandler {
